+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Unreleased
 ------------------
 Removed
  - Python 3.5 is no longer supported.
+ - The `tdl` module has been dropped.
 
 11.19.3 - 2021-01-07
 --------------------